seleph: improve performance scaling
For long runs the linear search in seleph is a significant performance
issue. Only ephemeris entries for which the time difference is within
tmax are considered and by pre-sorting the entries firstly by toe the
start index of the search can be found with a binary search. Further
the search can be exited early when beyond tmax.

This required changing the sorting order of nav->eph from ttr,toe,sat
to toe,ttr,sat, and in order to maintain exactly the same results
seleph() now explicitly checks and uses the most recently transmitted
entry when there a multiple entries with same toe.

Respective changes are made to selgeph() and selseph().