time2str: limit the output buffer size to 40 characters
Also have time2str() return a pointer to the output buffer, so that it
can be conveniently used as an function argument to trace and printf.

Normalize uses of the output buffer size to 40 characters which is
adequate for any precision output from time2str.

Similarly, the output buffer size of rinex.c:timestr_rnx() is limited
to 32 bytes, and the output buffer size of plotinfo.cpp:TimeStr() is
limited to 48 bytes, and uses updated.