rtcm3: implement lock time slip checking
Support determining the reported lock time from the lock time
indicators. Calculate the time since the last observation (per
signal). Use these to also check for a possible loss of lock. A slip
might now be reported on an outage.

Special case for receivers (Septentrio) that send a lock time
indicator of zero on a half cycle ambiguity. Consider these as an
outage deferring the decision to flag a slip until the signal
returns. For a short period of half cycle ambiguity that recovers, a
slip report might now be avoided.

The functions for converting the lock time indicator to a time also
support returning the time increment to the next indicator value. This
was not found to be needed.

Fix the observation flushing after a sync, which had been guarded by
the max obs overflow check so that once the max obs limit was reached
it was stuck.