sigindex: common function to resolve the frequence index
The code2idx() is not unique and conflicts need to be resolved. There
were different implementations across the code, RINEX, RTCM, the raw
receiver decoders. Add a common exported function to implement this
resolution.

Apply this resolution per satellite as the set of signals can vary
between satellites.

Update the Unicore and Novatel raw format decoder to use this new
common function.