rtksvr: implement automated antenna selection, pos in static mode
Implement the automated antenna type selection along with the antenna
delta from the input stream, from an RTCM or raw stream, the config
options ant{1,2}-anttype="*".

Have the config options ant{1,2}-postype RTCM and raw automatically
set the position, and independently of the antenna type, and now
including for the rover when in fixed mode.

Previously only an antenna delta would be picked up as an offset to a
base position, including a RTCM height, and only when the position was
also being automatically set and only for the base and not the rover.

Also there was no separate 'raw' antenna automated position mode, this
is included under the 'rtcm' option.