elevation mask pattern support
There was some support in the rtkplot code for elevation mask
patterns. Extended this support to the rtk and post processing
code. Move the code for the loading and saving of these patterns into
the main library and extend this to include a station name key so that
a single file can include the elevation mask patterns for the rover
and base, and add the file-elmaskfile option for this file. For post
processing, the station name can come from the RINEX files, from the
marker name. For rtkrcv, add new options sta1-name and sta2-name to
supply the rover and base station names respectively. These station
names are also need for the loading of DCB file, and would also be
useful for loading ocean tide loading parameters.

There was a '-sta' option to rtkrcv for reading the DCB file, but with
support for just the one station name this was not adequate for
reading the DCB for multiple receivers. The station name can nos be
specified in the config file using sta1-name and sta2-name so just
remove this option.