Misc issues with GLONASS not enabled in a build
rtcm3 save_msm_obs: better guard against a GLONASS satellite not being
found.

Avoid the nav geph array allocation when GLONASS is not enabled. It
was calling malloc() with a zero size allocation which may correctly
return NULL in this case, but the caller was using NULL to detected an
allocation failure and aborting.